Handle extra character in Slurm memory numbers that isn't a valid suffix.
For example: 0n should return 0.

I opened up an issue with SchedMD to get clarification on the sacct output
format.
In particular, the 'c' and 'n' suffixes on the ReqMem field.
I did some searching and it appears that these specify whether the request is
per core or per job.
This was not being handled correctly.
Move the 2nd suffix character out of MemoryUtils.py because it is Slurm
specific.

If the memory request is 0 then use MaxRSS + 10% for the memory request.

Don't flag CANCELLED jobs that didn't start as errors.

Add slurm tests for multi-node and array
While testing found issues with how Slurm reports memory and cores with
multiple nodes.
It reports it for the whole job, which makes sense, but parts of the code
were not taking this into account which lead to estimating the memory
per instance and cores per instance incorrectly.
The job memory and cores needs to be divided by the number of nodes when
selecting an instance type.

Resolves [Bug #17: Slurm parsing error](https://github.com/aws-samples/hpc-cost-simulator/issues/17)